+++ ReleaseNotes.md
@@ -30,6 +30,7 @@ SOFTWARE.
 * arm_chip_stm32: Added enum for last reset source (TODO: add more for sam & ino)
 * arm_inlines: added some binary manipulation inlines
 * arm_inlines: added nonnull check for pointers in SWAP_ENDxxB_TAB
+* arm_inlines: added temperature conversions (kelvin, celsius, fahrenheit)
 * arm_macros & arm_stdclib: simplified STR macro (no need of "")
 * arm_macros: moved TestMalloc to arm_macros.h (renamed to malloc_assert)
 * arm_macros: added OBJ macro to get the abstract object of a structure